DogHouse: Breaks out dog house discipline to dedicated .js file and hoisted up to commands
